Split function parsing out of `item.rs` to a new module.
`item.rs` is approaching the tidy-enforced length limit, and this will
make room for new additions.
